Rename BundleBuddyTypes.ts -> types.ts
Drop the "BundleBuddy" branding inherited from the upstream tool — the
types are no longer tied to it and the neutral name matches the
refreshed-PR2 layout. Pure rename; types unchanged.
