+++ build-tools/packages/build-cli/src/library/bundleSize/compareBundles.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import type { BundleComparison, BundleSummaries } from "./BundleBuddyTypes.js";
+import type { BundleComparison, BundleSummaries } from "./types.js";
 
 /**
  * Compares all the bundle summaries for a "baseline" and a "compare" bundle.
