+++ build-tools/packages/build-cli/src/library/bundleSize/index.ts
@@ -15,13 +15,13 @@ export {
 	IADOConstants,
 	SizeComparison,
 } from "./ADO/index.js";
+export { bundlesContainNoChanges, compareBundles } from "./compareBundles.js";
 export {
 	BundleComparison,
 	BundleMetric,
 	BundleMetricSet,
 	BundleSummaries,
-} from "./BundleBuddyTypes.js";
-export { bundlesContainNoChanges, compareBundles } from "./compareBundles.js";
+} from "./types.js";
 export {
 	GetBuildOptions,
 	getAllFilesInDirectory,
